requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
psycopg[binary]==3.2.12
python-dotenv==1.0.0
pandas==2.2.0
openpyxl==3.1.2
xlrd==2.0.1
Werkzeug==3.0.1
gunicorn==21.2.0
